fix authentication flow after Twitter API changes
Recently, Twitter made several API changes that broke
Net::Twitter, Net::Twitter::Lite and, by association, this module.
This, together with the desire to merge both solutions, led their
author to stop maintaining them and create a new distribution,
Twitter::API, which offered the current endpoints and a stable,
leaner and clearer interface. The only drawback is that it requires
perl 5.14+.

Because this module is no longer working with the old engines, this
patch makes the necessary updates to use Twitter::API instead. It
follows the orientations provided in Twitter::API::Trait::Migration
but does not use the Migration trait, since it is marked for
eventual removal.

https://developer.twitter.com/en/docs/basics/authentication/overview
https://developer.twitter.com/en/docs/twitter-for-websites/log-in-with-twitter/guides/implementing-sign-in-with-twitter
https://metacpan.org/pod/Net::Twitter#DESCRIPTION
https://metacpan.org/pod/Twitter::API#Migration-from-Net::Twitter-and-Net::Twitter::Lite
https://metacpan.org/pod/Twitter::API::Trait::Migration
https://metacpan.org/source/MMIMS/Twitter-API-1.0002/examples/oauth_web.psgi